Criada lógica de edição de worker permissions e unit tests de service permissions

+++ src/main/java/com/academy/services/ServiceService.java
@@ -13,12 +13,17 @@ import com.academy.models.service.Service;
 import com.academy.models.service.service_provider.ProviderPermissionEnum;
 import com.academy.models.service.service_provider.ServiceProvider;
 import com.academy.exceptions.EntityNotFoundException;
-import com.academy.models.ServiceType;
 import com.academy.models.Tag;
 import com.academy.repositories.ServiceRepository;
 import com.academy.repositories.ServiceTypeRepository;
 import com.academy.repositories.TagRepository;
+import com.academy.utils.Utils;
 import jakarta.transaction.Transactional;
+import org.apache.coyote.BadRequestException;
+import org.springframework.transaction.PlatformTransactionManager;
+import org.springframework.transaction.TransactionDefinition;
+import org.springframework.transaction.TransactionStatus;
+import org.springframework.transaction.support.DefaultTransactionDefinition;
 
 import java.util.Arrays;
 import java.util.ArrayList;
@@ -36,6 +41,7 @@ public class ServiceService {
     private final MemberService memberService;
     private final TagService tagService;
     private final ServiceTypeRepository serviceTypeRepository;
+    private final PlatformTransactionManager transactionManager;
     public ServiceService(ServiceRepository serviceRepository,
                           TagRepository tagRepository,
                           ServiceMapper serviceMapper,
@@ -43,7 +49,8 @@ public class ServiceService {
                           AuthenticationFacade authenticationFacade,
                           MemberService memberService,
                           TagService tagService,
-                          ServiceTypeRepository serviceTypeRepository) {
+                          ServiceTypeRepository serviceTypeRepository,
+                          PlatformTransactionManager transactionManager) {
         this.serviceRepository = serviceRepository;
         this.tagRepository = tagRepository;
         this.serviceMapper = serviceMapper;
@@ -52,11 +59,12 @@ public class ServiceService {
         this.memberService = memberService;
         this.tagService = tagService;
         this.serviceTypeRepository = serviceTypeRepository;
+        this.transactionManager = transactionManager;
     }
 
     // Create
     @Transactional
-    public ServiceResponseDTO create(ServiceRequestDTO dto) {
+    public ServiceResponseDTO create(ServiceRequestDTO dto) throws AuthenticationException, BadRequestException {
         Member member = memberService.getMemberByUsername(authenticationFacade.getUsername());
         Service service = serviceMapper.toEntity(dto, member.getId());
 
@@ -65,26 +73,20 @@ public class ServiceService {
 
         Service savedService = serviceRepository.save(service);
 
-        createOwnerServiceProvider(new ServiceProviderRequestDTO(
-                member.getId(),
-                savedService.getId(),
-                Arrays.asList(ProviderPermissionEnum.values())
-        ));
+        createOwnerServiceProvider(member.getId(), service.getId());
         return serviceMapper.toDto(savedService, getPermissionsByProviderUsernameAndServiceId(member.getUsername(), savedService.getId()));
     }
-    private ServiceProviderResponseDTO createOwnerServiceProvider(ServiceProviderRequestDTO request) {
-        return serviceProviderService.createServiceProvider(request);
-    }
+
 
     // Update
     @Transactional
-    public ServiceResponseDTO update(Long id, ServiceRequestDTO dto) {
+    public ServiceResponseDTO update(Long id, ServiceRequestDTO dto) throws AuthenticationException{
         String username = authenticationFacade.getUsername();
         Service existing = serviceRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException(Service.class,id));
 
         List<ProviderPermissionEnum> permissions = getPermissionsByProviderUsernameAndServiceId(username, existing.getId());
-        if(permissions == null || !permissions.contains(ProviderPermissionEnum.UPDATE))
+        if(!checkIfHasPermission(permissions,ProviderPermissionEnum.UPDATE))
             throw new AuthenticationException("Member doesn't have permission to update service");
 
         // Remove existing tag associations
@@ -115,13 +117,13 @@ public class ServiceService {
 
     // Read one
     public ServiceResponseDTO getById(Long id) {
-        Service service = serviceRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException(Service.class,id));
+        Service service = getEntityById(id);
         String username =  authenticationFacade.getUsername();
-        List<ProviderPermissionEnum> permissions = getPermissionsByProviderUsernameAndServiceId(username, id);
-        if(permissions == null || !permissions.contains(ProviderPermissionEnum.READ))
-            throw new AuthenticationException("Member doesn't have permission to read service");
         return serviceMapper.toDto(service, getPermissionsByProviderUsernameAndServiceId(username, service.getId()));
+    }
+    public Service getEntityById(Long id){
+        return serviceRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException(Service.class,id));
     }
 
     // Delete
@@ -131,7 +133,7 @@ public class ServiceService {
         Service service = serviceRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException(Service.class, id));
         List<ProviderPermissionEnum> permissions = getPermissionsByProviderUsernameAndServiceId(username, id);
-        if(permissions == null || !permissions.contains(ProviderPermissionEnum.DELETE))
+        if(!checkIfHasPermission(permissions,ProviderPermissionEnum.DELETE))
             throw new AuthenticationException("Member doesn't have permission to delete service");
 
         service.removeAllTags();
@@ -147,21 +149,70 @@ public class ServiceService {
 
     public List<ProviderPermissionEnum> getPermissionsByProviderUsernameAndServiceId(String username, Long serviceId){
         if(!hasServiceProvider(username, serviceId))
-            return null;
+            return new ArrayList<ProviderPermissionEnum>();
         return serviceProviderService.getPermissionsByProviderUsernameAndServiceId(username, serviceId);
+    }
+    //TODO ver isto que nao esta a ser usado
+    public List<ProviderPermissionEnum> getPermissionsByProviderIdAndServiceId(Long id, Long serviceId){
+        if(!hasServiceProvider(id, serviceId))
+            return new ArrayList<ProviderPermissionEnum>();
+        return serviceProviderService.getPermissionsByProviderIdAndServiceId(id, serviceId);
     }
     private boolean hasServiceProvider(String username, Long serviceId){
         return serviceProviderService.existsByServiceIdAndProviderUsername(serviceId, username);
     }
-    /*
+    private boolean hasServiceProvider(Long id, Long  serviceId){
+        return serviceProviderService.existsByServiceIdAndProviderId(id, serviceId);
+    }
+
     @Transactional
-    public ServiceResponseDTO updateMemberPermissions(Long serviceId, Long memberId, List<ProviderPermissionEnum> permissions){
-        ServiceProvider serviceProvider = serviceProviderService.getByServiceIdAndMemberId(serviceId, memberId);
+    public ServiceResponseDTO updateMemberPermissions(Long serviceId, Long memberToBeUpdatedId, List<ProviderPermissionEnum> newPermissions) throws AuthenticationException, BadRequestException {
+        String updaterUsername =   authenticationFacade.getUsername();
+        Long updaterId = memberService.getMemberByUsername(updaterUsername).getId();
+        ServiceProvider serviceProvider;
+        try{
+            serviceProvider = serviceProviderService.getByServiceIdAndMemberId(serviceId, memberToBeUpdatedId);
+        }catch(EntityNotFoundException e){
+            throw new BadRequestException("The service with id " +  serviceId + " does not have a Service Provider with the id " +memberToBeUpdatedId);
+        }
+        Member memberToBeUpdated =  memberService.getMemberById(memberToBeUpdatedId);
+        List<ProviderPermissionEnum> oldPermissions = getPermissionsByProviderUsernameAndServiceId(memberToBeUpdated.getUsername(), serviceId);
+        List<ProviderPermissionEnum> updaterPermissions = getPermissionsByProviderUsernameAndServiceId(updaterUsername, serviceId);
+
+        validateUpdateOfPermissions(updaterPermissions, oldPermissions, newPermissions, updaterId, memberToBeUpdatedId);
+
         serviceProviderService.deleteAllPermissions(serviceProvider);
-        serviceProviderService.addPermissions(serviceProvider, permissions);
+        serviceProviderService.addPermissions(serviceProvider, newPermissions);
         return getById(serviceId);
     }
-    */
 
+    private ServiceProviderResponseDTO createOwnerServiceProvider(Long memberId, Long serviceId) throws AuthenticationException, BadRequestException {
+        return serviceProviderService.createServiceProvider(new ServiceProviderRequestDTO(
+                memberId,
+                serviceId,
+                Arrays.asList(ProviderPermissionEnum.values()),
+                true
+        ));
+    }
+    private void validateUpdateOfPermissions(List<ProviderPermissionEnum> updaterPermissions,
+                                             List<ProviderPermissionEnum> oldPermissions,
+                                             List<ProviderPermissionEnum> newPermissions,
+                                             Long updaterId,
+                                             Long memberToBeUpdatedId) throws BadRequestException {
+        boolean isOwnerBeingUpdated = oldPermissions.contains(ProviderPermissionEnum.OWNER);
+        if(!checkIfHasPermission(updaterPermissions, ProviderPermissionEnum.UPDATE_PERMISSIONS))
+                throw new AuthenticationException("Member does not have permission to update permissions");
+
+        if(updaterId == memberToBeUpdatedId)
+            throw new BadRequestException("Cannot edit your own permissions");
+        if(isOwnerBeingUpdated)
+            throw new BadRequestException("Cannot edit owner's permissions");
+        
+        ServiceProviderService.checkIfValidPermissions(newPermissions);
+    }
+
+    private boolean checkIfHasPermission(List<ProviderPermissionEnum> permissions, ProviderPermissionEnum permission ) {
+        return Utils.hasPermission(permissions, permission);
+    }
 
 }
